+++ sdk/resourcemanager/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/models/BgpPeerStatus.java
@@ -4,67 +4,73 @@
 
 package com.azure.resourcemanager.network.models;
 
+import com.azure.core.annotation.Fluent;
 import com.azure.core.annotation.Immutable;
 import com.azure.core.util.logging.ClientLogger;
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-/** BGP peer status details. */
+/**
+ * BGP peer status details.
+ */
 @Immutable
 public final class BgpPeerStatus {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(BgpPeerStatus.class);
+    @JsonIgnore
+    private final ClientLogger logger = new ClientLogger(BgpPeerStatus.class);
 
     /*
-     * The virtual network gateway's local address.
+     * The virtual network gateway's local address
      */
     @JsonProperty(value = "localAddress", access = JsonProperty.Access.WRITE_ONLY)
     private String localAddress;
 
     /*
-     * The remote BGP peer.
+     * The remote BGP peer
      */
     @JsonProperty(value = "neighbor", access = JsonProperty.Access.WRITE_ONLY)
     private String neighbor;
 
     /*
-     * The autonomous system number of the remote BGP peer.
+     * The autonomous system number of the remote BGP peer
      */
     @JsonProperty(value = "asn", access = JsonProperty.Access.WRITE_ONLY)
-    private Long asn;
+    private Integer asn;
 
     /*
-     * The BGP peer state.
+     * The BGP peer state
      */
     @JsonProperty(value = "state", access = JsonProperty.Access.WRITE_ONLY)
     private BgpPeerState state;
 
     /*
-     * For how long the peering has been up.
+     * For how long the peering has been up
      */
     @JsonProperty(value = "connectedDuration", access = JsonProperty.Access.WRITE_ONLY)
     private String connectedDuration;
 
     /*
-     * The number of routes learned from this peer.
+     * The number of routes learned from this peer
      */
     @JsonProperty(value = "routesReceived", access = JsonProperty.Access.WRITE_ONLY)
     private Long routesReceived;
 
     /*
-     * The number of BGP messages sent.
+     * The number of BGP messages sent
      */
     @JsonProperty(value = "messagesSent", access = JsonProperty.Access.WRITE_ONLY)
     private Long messagesSent;
 
     /*
-     * The number of BGP messages received.
+     * The number of BGP messages received
      */
     @JsonProperty(value = "messagesReceived", access = JsonProperty.Access.WRITE_ONLY)
     private Long messagesReceived;
 
     /**
-     * Get the localAddress property: The virtual network gateway's local address.
-     *
+     * Get the localAddress property: The virtual network gateway's local
+     * address.
+     * 
      * @return the localAddress value.
      */
     public String localAddress() {
@@ -73,7 +79,7 @@ public final class BgpPeerStatus {
 
     /**
      * Get the neighbor property: The remote BGP peer.
-     *
+     * 
      * @return the neighbor value.
      */
     public String neighbor() {
@@ -81,17 +87,18 @@ public final class BgpPeerStatus {
     }
 
     /**
-     * Get the asn property: The autonomous system number of the remote BGP peer.
-     *
+     * Get the asn property: The autonomous system number of the remote BGP
+     * peer.
+     * 
      * @return the asn value.
      */
-    public Long asn() {
+    public Integer asn() {
         return this.asn;
     }
 
     /**
      * Get the state property: The BGP peer state.
-     *
+     * 
      * @return the state value.
      */
     public BgpPeerState state() {
@@ -99,8 +106,9 @@ public final class BgpPeerStatus {
     }
 
     /**
-     * Get the connectedDuration property: For how long the peering has been up.
-     *
+     * Get the connectedDuration property: For how long the peering has been
+     * up.
+     * 
      * @return the connectedDuration value.
      */
     public String connectedDuration() {
@@ -108,8 +116,9 @@ public final class BgpPeerStatus {
     }
 
     /**
-     * Get the routesReceived property: The number of routes learned from this peer.
-     *
+     * Get the routesReceived property: The number of routes learned from this
+     * peer.
+     * 
      * @return the routesReceived value.
      */
     public Long routesReceived() {
@@ -118,7 +127,7 @@ public final class BgpPeerStatus {
 
     /**
      * Get the messagesSent property: The number of BGP messages sent.
-     *
+     * 
      * @return the messagesSent value.
      */
     public Long messagesSent() {
@@ -127,7 +136,7 @@ public final class BgpPeerStatus {
 
     /**
      * Get the messagesReceived property: The number of BGP messages received.
-     *
+     * 
      * @return the messagesReceived value.
      */
     public Long messagesReceived() {
@@ -136,7 +145,7 @@ public final class BgpPeerStatus {
 
     /**
      * Validates the instance.
-     *
+     * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {
